[FE] 홈 화면 제작완료
