+++ fe/src/app/Home.module.scss
@@ -1,73 +1,4 @@
-@use "../styles/mixins" as *;
-@use "../styles/variables" as *;
-
 .home-container {
   display: grid;
   grid-template-columns: min(40%, 500px) 1fr;
-}
-
-.info {
-  &-box {
-    position: sticky;
-    top: 0;
-    left: 0;
-    z-index: 100;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    width: 100%;
-    height: 100vh;
-    background-color: $color-beige-100;
-    padding: 0 1.25rem;
-    border-right: 1px solid $color-green-50;
-  }
-
-  &-item {
-    max-width: 370px;
-
-    &.fade-up {
-      animation: fadeUp 0.5s ease forwards;
-    }
-
-    @keyframes fadeUp {
-      0% {
-        opacity: 0;
-        transform: translateY(30px);
-      }
-      100% {
-        opacity: 1;
-        transform: translateY(0);
-      }
-    }
-  }
-
-  &-title {
-    max-width: 360px;
-    @include myeongjo($font-size-50, $font-weight-bold);
-    line-height: 1.1;
-    color: $color-green-500;
-    margin-bottom: 0.825rem;
-  }
-
-  &-subtitle {
-    font-size: $font-size-20;
-    color: $color-green-500;
-    margin-bottom: 1.5rem;
-  }
-
-  &-content {
-    font-size: $font-size-18;
-    word-break: keep-all;
-    margin-bottom: 2rem;
-  }
-}
-
-.section {
-  &-item {
-    position: relative;
-    height: 100vh;
-  }
 }
