+++ fe/src/components/Home/ImageSection/ImageSection.module.scss
@@ -1,0 +1,6 @@
+.section {
+  &-item {
+    position: relative;
+    height: 100vh;
+  }
+}
